pyproject.toml
[tool.poetry]
name = "Moebot"
version = "1.0.0"
description = "A Reddit to Discord/Twitter bot"
authors = ["glitchy <thepatheticweebgamer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.10.0"
python-dotenv = "^0.20.0"
httpx = "^0.23.0"
loguru = "^0.6.0"
pygelbooru = "^0.5.0"
discord-webhook = {extras = ["async"], version = "^0.17.0"}

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
